Remove Legacy OTP Secret code (#34743)
