+++ db/post_migrate/20250520192024_remove_legacy_devise_two_factor_secrets_from_users.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class RemoveLegacyDeviseTwoFactorSecretsFromUsers < ActiveRecord::Migration[7.1]
+  def change
+    safety_assured do
+      remove_column :users, :encrypted_otp_secret, :string
+      remove_column :users, :encrypted_otp_secret_iv, :string
+      remove_column :users, :encrypted_otp_secret_salt, :string
+    end
+  end
+end
